initial commit
To perform the task with an ease, a pre-trained neural network is used with the pretrained parameters i.e. VGG19. The Convolutional Neural Network is built to classify various breeds of Cheetah as in this scenario, the network is organized with some layers and normalizers to classify the images(present in Cheetah, Hyena, Jaguar and Tiger dataset) into their specified classes as Cheetah, Hyena, Jaguar and Tiger. VGG19 is used to attain better results with highly trained parameters and dense network which is a perfect arrangement of Max pooling layers and Convolutional layers.